backend/requirements.txt

flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymysql==1.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.1
gunicorn==21.2.0
python-dotenv==1.0.0
